pyproject.toml
[tool.poetry]
name = "clandestino"
version = "0.1.0"
description = ""
authors = ["XimitGaia <im.ximit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"

clandestino-interfaces = {path = "/home/ximit/Projects/clandestino/extras/interfaces/dist/clandestino_interfaces-0.1.0.tar.gz"}
clandestino-postgres = {path = "/home/ximit/Projects/clandestino/extras/postgres/dist/clandestino_postgres-0.1.0.tar.gz"}
clandestino-mongo = {path = "/home/ximit/Projects/clandestino/extras/mongo/dist/clandestino_mongo-0.1.0.tar.gz"}

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cdt = "clandestino.clandestino:main"